Add poetry.lock file to the project
The poetry.lock file has been added, as a means of locking the project's dependencies. This includes the specifications for all of the packages, their versions, and their dependencies. Using the poetry.lock file allows for much more control over the specific versions of packages used in the project. The added file includes details for packages like aiohttp, aiosignal, annotated-types, anyio, asttokens, attrs, beautifulsoup4, and certifi.